Tree: fix _insertNodesInDOM
If nodes are added to a tree that currently does not have any visible
nodes (e.g. due to a filter), these nodes need to be rendered (if
visible). Therefore, check if the new node will appear inside the view
range. One part of this check is if it is possible to render any node,
i.e. is the size of the viewRange larger than 0. To achieve the correct
result check "this.viewRangeSize > 0" instead of
"this.viewRangeRendered.size() > 0", otherwise nodes added to an "empty"
tree are never rendered.

310065